chore: fix all clippy warnings across workspace
Fixed 268 clippy warnings across all crates:

**Mechanical fixes:**
- Replaced useless format!() with String methods
- Removed unnecessary closures and redundant casts
- Used .is_multiple_of() instead of manual % checks
- Removed .clone() on Copy types
- Used .is_some_and() instead of .map_or(false, ...)
- Replaced vec![] + push patterns with vec![...] literals
- Used .first() instead of .get(0)
- Removed unnecessary & dereferences

**Suppressed architectural warnings:**
- Added #[allow(clippy::too_many_arguments)] on complex functions
- Added #[allow(clippy::arc_with_non_send_sync)] on single-threaded Arc usage
- Added #[allow(clippy::mut_from_ref)] on interior mutability pattern

**Test fixes:**
- Fixed engine-authoring test compilation (missing Scene/Sprite fields)
- Fixed field_metadata_to_schema import in schema builders tests

**Derived Default impls:**
- Added #[derive(Default)] to 20+ structs in rust-os
- Added Default to AnimationDispatcher, AssetCache, PipelineStrategies

Verification:
- cargo clippy --workspace → 0 warnings
- cargo test -p engine-core -p engine-authoring → all pass
- cargo run --check-scenes → asteroids validates cleanly

+++ engine/src/game_loop.rs
@@ -246,8 +246,8 @@ pub fn game_loop(
                 if let Some(buf) = world.output_buffer() {
                     (
                         buf.last_diff_count,
-                        buf.dirty_cell_count() as u32,
-                        buf.total_cells() as u32,
+                        buf.dirty_cell_count(),
+                        buf.total_cells(),
                         buf.write_count,
                     )
                 } else {
